doing some fixes

+++ utils/redis.js
@@ -1,38 +1,60 @@
 import { createClient } from 'redis';
-import { promisify } from 'util';
 
 class RedisClient {
   constructor() {
+
     this.client = createClient();
-    this.getAsync = promisify(this.client.get).bind(this.client);
-
-    this.client.on('error', (error) => {
-      console.log(`Redis client not connected to the server: ${error.message}`);
+    this.ready = true;
+    this.client.on('error', (err) => {
+      console.log('Redis Client Error', err);
+      this.ready = false;
     });
-
-    this.client.on('connect', () => {
-      // console.log('Redis client connected to the server');
-    });
+    this.client.connect();
   }
 
   isAlive() {
-    return this.client.connected;
+    return this.ready;
   }
 
   async get(key) {
-    const value = await this.getAsync(key);
-    return value;
+    if (!key) {
+      console.error('Usage: get(key)');
+      return null;
+    }
+    try {
+      return await this.client.get(key);
+    } catch (err) {
+      console.log('Error getting value', err);
+      return null;
+    }
   }
 
   async set(key, value, duration) {
-    this.client.setex(key, duration, value);
+    if (!key || !value || !duration) {
+      console.error('Usage: set(key, value, duration)');
+      return;
+    }
+    try {
+      await this.client.set(key, value, {
+        EX: duration,
+      });
+    } catch (err) {
+      console.log('Error setting key', err);
+    }
   }
 
   async del(key) {
-    this.client.del(key);
+    if (!key) {
+      console.error('Usage: del(key)');
+      return;
+    }
+    try {
+      await this.client.del(key);
+    } catch (err) {
+      console.log('Error deleting key', err);
+    }
   }
 }
 
 const redisClient = new RedisClient();
-
 export default redisClient;
